% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPT.R
\docType{package}
\name{NPT-package}
\alias{NPT-package}
\alias{NPT}
\title{Non-Parametric Transformation (NPT) Model}
\description{
Functions written in C++ to fit non-parametric transformation (NPT) model.
}
\keyword{package}
