% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPT.R
\docType{package}
\name{NPT-package}
\alias{NPT-package}
\alias{NPT}
\title{Non-Parametric Transformation Model}
\description{
Implements the simplex algorithm for fitting non-parametric transformation models. The computationally intensive part is written in C++.
}
\keyword{package}
